+++ src/main/java/neresources/api/distributions/DistributionUnderWater.java
@@ -4,7 +4,7 @@ import neresources.api.utils.DistributionHelpers;
 
 public class DistributionUnderWater extends DistributionBase{
 
-    public DistributionUnderWater(double maxChance) {
+    public DistributionUnderWater(float maxChance) {
         super(DistributionHelpers.getUnderwaterDistribution(maxChance));
         this.bestHeight = 61;
     }
